+++ apps/webapp/charmClient/apis/spacesApi.ts
@@ -3,8 +3,6 @@ import type { Space, Prisma, RewardsGithubRepo } from '@charmverse/core/prisma';
 import * as http from '@packages/adapters/http';
 import type { DowngradeSubscriptionTierRequest } from '@packages/lib/subscription/downgradeSubscriptionTier';
 import type { UpgradeSubscriptionTierRequest } from '@packages/lib/subscription/upgradeSubscriptionTier';
-import type { CreateSpaceContributionRequest } from '@packages/spaces/createSpaceContribution';
-import type { SpaceReceipt } from '@packages/spaces/getSpaceReceipts';
 
 import type { CreateSpaceProps } from 'lib/spaces/createSpace';
 import type { CustomDomainVerification } from 'lib/spaces/interfaces';
@@ -124,24 +122,12 @@ export class SpacesApi {
     return http.PUT<RewardsGithubRepo>(`/api/spaces/${spaceId}/github/repo/${repoId}`, payload);
   }
 
-  createSpaceContribution(spaceId: string, payload: CreateSpaceContributionRequest) {
-    return http.POST(`/api/spaces/${spaceId}/space-contribution`, payload);
-  }
-
-  getSpaceContributions(spaceId: string) {
-    return http.GET<SpaceReceipt[]>(`/api/spaces/${spaceId}/space-contribution`);
-  }
-
   getSpaceTokenBalance(spaceId: string) {
-    return http.GET<number>(`/api/spaces/${spaceId}/space-token-balance`);
+    return http.GET<number>(`/api/spaces/${spaceId}/token-balance`);
   }
 
   upgradeSubscriptionTier(spaceId: string, payload: UpgradeSubscriptionTierRequest) {
     return http.POST<void>(`/api/spaces/${spaceId}/upgrade-tier`, payload);
-  }
-
-  cancelSubscriptionTier(spaceId: string) {
-    return http.POST<void>(`/api/spaces/${spaceId}/cancel-tier`);
   }
 
   downgradeSubscriptionTier(spaceId: string, payload: DowngradeSubscriptionTierRequest) {
