+++ packages/lib/src/subscription/chargeSpaceSubscription.ts
@@ -13,8 +13,7 @@ export const SubscriptionTierAmountRecord: Record<SpaceSubscriptionTier, number>
   silver: 25,
   gold: 100,
   grant: 0,
-  readonly: 0,
-  cancelled: 0
+  readonly: 0
 };
 
 export async function chargeSpaceSubscription({ spaceId }: { spaceId: string }) {
@@ -32,10 +31,6 @@ export async function chargeSpaceSubscription({ spaceId }: { spaceId: string }) 
     throw new Error('Space is not a subscription space');
   }
 
-  if (space.subscriptionTier === 'cancelled') {
-    throw new Error('Space subscription is cancelled');
-  }
-
   if (!UpgradableTiers.includes(space.subscriptionTier as UpgradableTier)) {
     throw new Error('Space subscription is not chargeable');
   }
@@ -49,12 +44,22 @@ export async function chargeSpaceSubscription({ spaceId }: { spaceId: string }) 
   const subscriptionTierAmountInWei = parseUnits(subscriptionTierAmount.toString(), 18);
 
   if (spaceTokenBalanceInWei < subscriptionTierAmountInWei) {
-    await prisma.space.update({
-      where: { id: spaceId },
-      data: {
-        subscriptionTier: 'readonly'
-      }
-    });
+    await prisma.$transaction([
+      prisma.space.update({
+        where: { id: spaceId },
+        data: {
+          subscriptionTier: 'readonly'
+        }
+      }),
+      prisma.spaceSubscriptionTierChangeEvent.create({
+        data: {
+          spaceId,
+          newTier: 'readonly',
+          previousTier: subscriptionTier
+        }
+      })
+    ]);
+
     log.warn(`Insufficient space token balance, space downgraded to free tier`, {
       spaceId,
       spaceTokenBalance,
